requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
imageio==2.37.0
kiwisolver==1.4.8
lazy_loader==0.4
matplotlib==3.10.0
networkx==3.4.2
numpy<2
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
scikit-image==0.25.1
scipy==1.15.1
seaborn==0.13.2
six==1.17.0
tifffile==2025.1.10
tqdm==4.67.1
tzdata==2025.1
flask
gunicorn
torch
torchvision
tensorflow
